[CHANGE] マイグレーション関連のクラスをマイグレーションクラスに移動
